pyproject.toml
[project]
name = "deeplab"
version = "0.1.0"
description = "基于Agent的AI科研辅助平台"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "duckdb>=1.4.4",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.0",
    "fastembed>=0.7.4",
    "google-genai>=1.65.0",
    "pypdf>=5.5.0",
    "pytz>=2025.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
